Add more WP-CLI functionality to the blueprints
Add install options from WP-CLI to the blueprint functionality of VV.
Now includes `—version`, `—force`, `—activate` and `—activate_network`.

Also changed `is_bool()` in `vv-install` to a string comparison since
`is_bool()` does not work when the boolean you’re trying to pass is
actually a string.

Included PHPDoc comment blocks for readability and future reference as
well as comments throughout the script to make it a little easier to
understand what is going on.