Rewrite how vv-install handles a theme or plugin being installed using the user/repository format
Rewrite how `vv-install` handles a theme or plugin being installed
using the `user/repository` format.

Now downloads the tarball of the repository, unzips that, moves it into
the proper folder renaming it so the folder name is equal to the
repository name e.g. `vv-master` becomes `vv` and then activates the
plugin or theme based on the blueprint settings for this specific
object.