Support WP Multi-Network configuration from blueprint installs.
This commit adds recognition for a `networks` key in a blueprint and
a `network_domain` key in a subsite definition. It will automatically
network-activate the WP Multi-Network plugin if this is found and then
create the networks defined in the `networks` key. Subsites that have
either a `network_id` or `network_domain` key in their blueprint will
subsequently be added to the appropriate network.