Add WP Multi-Network domain configuration from a blueprint file.
Similar to a single WP Multisite's subdomains defined in a blueprint,
this commit recognizes a special blueprint-level root key named like
`$BLUEPRINT::subnetwork_domains` and writes a union of the subdomains
and the subnetwork domains to the `vvv-hosts` file in time for the
`vagrant-hostsupdater` plugin to pick them up during a `vagrant up`.
Also includes the new root domains (and wildcard subdomain) in the nginx
configuration file to correctly resolve multihomed domain to WP install.